backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo[srv]==4.6.0
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
firebase-admin==6.2.0
anthropic==0.8.1
google-cloud-vision==3.5.0
python-jose[cryptography]==3.3.0
schedule==1.2.0
Pillow==10.1.0
requests==2.31.0
